Add new option -bg to hide console window and run in background.
This can be used to workaround issue in current WSL preview, where (unlike previous
versions) executing a Windows console application from within WSL opens it in a new
console window.

Note that as long as the application is built as a console application, it will still
start with a console, but when option -bg is specified it will hide it immediately.
You may still notice the console window flashing briefly at startup. To avoid this
the application can be built as a GUI application, by adding to go build
-ldflags "-H windowsgui".