port vmserial-connect to powershell
in case you want to enter the serial console on a Hyper-V server
_without_ WSL

still requires `socat`, obviously, but obtaining it on Windows is a
little more complicated than using your package manager. I've tested
with the one from
[unix-utils](https://sourceforge.net/projects/unix-utils/files/socat/1.7.3.2/)
and it works great, but if I install msys2 and install socat that way it
doesn't work.

I would prefer an easier way to put socat in, perhaps a
statically-compiled one (or a wild enough wish could even be just to
implement the fancy PTY stuff necessary in npiperelay itself?)